personal_website/requirements.txt
streamlit
plotly
pandas
pillow
numpy
requests